FIX cake controllers pending update img

+++ src/api/controllers/cake.js
@@ -4,12 +4,14 @@ const Cake = require("../models/cake");
 const createCake = async (req, res, next) => {
     try {
         const { name, difficulty } = req.body;
+
         if (difficulty !== "Baja" && difficulty !== "Media" && difficulty !== "Alta") {
-            return res.status(400).json({ message: 'La dificultad ha sido mal introducida. Introduce: Baja, Media o Alta' })
+            return res.status(400).json({ message: 'La dificultad no ha sido introducida o ha sido mal introducida. Introduce: Baja, Media o Alta' })
         };
+
         const cakeDuplicated = await Cake.findOne({ name });
         if (cakeDuplicated) {
-            return res.status(400).json({ message: `La tarta ${name} ya ha sido creada.` })
+            return res.status(400).json({ message: `La tarta ${name} ya ha sido creada anteriormente.` })
         };
         
         const newCake = new Cake(req.body);
@@ -56,7 +58,7 @@ const getCakeByName = async (req, res, next) => {
         if (!findCakeByName.length) {
             return res.status(400).json({ message: `No existe ninguna tarta en nuestro stock que contenga '${name}'` });
         }
-        return res.status(200).json({ message: "Tarta encontrada:", findCakeByName })
+        return res.status(200).json({ message: `Con la letra ${name} tenemos las siguientes tartas:`, findCakeByName })
     } catch (error) {
         return res.status(400).json(`❌ Fallo en getCakeByName: ${error.message}`)
     }
@@ -65,18 +67,62 @@ const getCakeByName = async (req, res, next) => {
 const updateCake = async (req, res, next) => {
     try {
         const { id } = req.params;
-        const cakeModify = new Cake(req.body)
+        const { ingredients } = req.body;
+
+        const cakeModify = new Cake(req.body);
         cakeModify._id = id;
-        if (req.files) {
-            cakeModify.img = req.files.path
-            const oldCake = await Cake.findById(id);
-            deleteFile(oldCake.img);
-        }
+        if (cakeModify) {
+            cakeModify.$addToSet = { ingredients }
+        };
+
+        if (cakeModify.difficulty !== "Baja" && cakeModify.difficulty !== "Media" && cakeModify.difficulty !== "Alta" && cakeModify.difficulty !== undefined) {
+            return res.status(400).json({ message: 'La dificultad ha sido mal introducida. Introduce: Baja, Media o Alta' })
+        };
+        // const restParams = { ...rest };
+        // if (ingredients) {
+        //     restParams.$addToSet = { ingredients }
+        // }
+
+
+
+        // const cakeModify = new Cake(req.body)
+        // cakeModify._id = id;
+        // if (req.files) {
+        //     cakeModify.img = req.files.path
+        //     const oldCake = await Cake.findById(id);
+        //     if (oldCake) {
+        //         if (oldCake.img) {
+        //             deleteFile(oldCake.img);
+        //         }
+        //         if (oldCake.img2) {
+        //             deleteFile(oldCake.img2);
+        //         }
+        //         if (oldCake.img3) {
+        //             deleteFile(oldCake.img3);
+        //         }
+        //     }
+        // }
+
+        // if (req.files) {
+        //     const oldCake = await Cake.findById(id);
+        //     if (oldCake) {
+        //         if (oldCake.img) {
+        //             deleteFile(oldCake.img);
+        //         }
+        //         if (oldCake.img2) {
+        //             deleteFile(oldCake.img2);
+        //         }
+        //         if (oldCake.img3) {
+        //             deleteFile(oldCake.img3);
+        //         }
+        //     }
+        // }
 
         const cakeUpdated = await Cake.findByIdAndUpdate(id, cakeModify, { new: true });
         if (!cakeUpdated){
             return res.status(400).json({ message: 'No existe esta tarta.' });
         }
+        
         return res.status(200).json({ message: 'Tarta actualizada correctamente', cakeUpdated });
     } catch (error) {
         return res.status(400).json(`❌ Fallo en updateCake: ${error.message}`)
@@ -90,7 +136,17 @@ const deleteCake = async (req, res, next) => {
         if (!cakeDeleted) {
             return res.status(400).json({ message: "Esta tarta ya no existe en nuestro stock." });
         }
-        deleteFile(cakeDeleted.img);
+        if (cakeDeleted) {
+            if (cakeDeleted.img) {
+                deleteFile(cakeDeleted.img);
+            }
+            if (cakeDeleted.img2) {
+                deleteFile(cakeDeleted.img2);
+            }
+            if (cakeDeleted.img3) {
+                deleteFile(cakeDeleted.img3);
+            }
+        }
         return res.status(200).json({ message: `Tarta eliminada de nuestro stock.`, cakeDeleted });
     } catch (error) {
         return res.status(400).json(`❌ Fallo en deleteCake: ${error.message}`)
